[OrderStorage] add orders_storage_operations
